requirements-dev.frozen.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements-dev.frozen.txt requirements.in requirements-chrouter.in requirements-dev.in requirements-doc.in requirements-tests.in
#
-e git+https://github.com/assembl/Beaker.git@assembl#egg=Beaker
-e git+https://github.com/maparent/cornice.git@master#egg=cornice
-e git+https://github.com/pythonforfacebook/facebook-sdk.git@ddc911f54d6136e149c03d145fd3760751a7d50d#egg=facebook_sdk
-e git+https://github.com/ecreall/graphql-wsgi.git@090df73f0184f20a643c3d3f413e55f3f4be3d45#egg=graphql-wsgi
-e git+https://github.com/dgilland/hashfs.git@386ce3b8c69576007549ddf5122282099b87fe42#egg=hashfs
-e git+https://github.com/assembl/jasmine-splinter-runner.git#egg=jasmine-splinter-runner
-e git+https://github.com/benoitg/jwzthreading.git@master#egg=jwzthreading
-e git+https://github.com/maparent/langdetect.git@no_sample#egg=langdetect
-e git+https://github.com/assembl/pyramid_beaker.git@assembl#egg=pyramid_beaker
-e git+https://github.com/assembl/sockjs-tornado.git@conflicting_localhost#egg=sockjs_tornado
-e git+https://github.com/assembl/sqltap.git@assembl#egg=sqltap
alabaster==0.7.10         # via sphinx
alembic==0.8.6
amqp==2.3.2               # via kombu
anyjson==0.3.3
appnope==0.1.0            # via ipython
asn1crypto==0.24.0        # via cryptography
aspy.yaml==1.0.0          # via pre-commit
babel==2.3.4
backports-abc==0.5        # via tornado
backports.shutil-get-terminal-size==1.0.0  # via ipython
backports.ssl-match-hostname==3.5.0.1  # via tornado
bcrypt==3.1.4             # via paramiko
beautifulsoup4==4.6.0
billiard==3.5.0.3         # via celery
cached-property==1.4.0    # via pre-commit
celery==4.1.1
certifi==2018.1.18        # via tornado
cffi==1.11.5              # via bcrypt, cryptography, pynacl
cfgv==1.0.0               # via pre-commit
chameleon==3.2            # via lingua
characteristic==14.3.0    # via montague
click==6.7                # via pip-tools
colander==1.3.1
colanderalchemy==0.3.3
colorama==0.3.9
coloredlogs==7.3
configparser==3.5.0       # via flake8
contextlib2==0.5.5        # via raven
cov-core==1.15.0
coverage==4.1
cryptography==2.1.4
cssselect==1.0.3          # via premailer
cssutils==1.0.2           # via premailer
decorator==4.2.1          # via ipython, jsonpath-ng, traitlets
defusedxml==0.4.1         # via python-saml
dm.xmlsec.binding==1.3.2  # via python-saml
dnspython==1.15.0         # via pyisemail
docutils==0.14            # via sphinx
dogpile.cache==0.6.7
elasticsearch==5.1.0
enum34==1.1.6
et-xmlfile==1.0.1         # via openpyxl
executor==19.1            # via rotate-backups
fabric==1.13.2
fasteners==0.14.1         # via executor
feedparser==5.2.1
first==2.0.1              # via pip-tools
flake8==3.5.0
flaky==3.1.1
funcsigs==1.0.2           # via mock
future==0.15.2
gnureadline==6.3.8        # via ipython
google-api-python-client==1.6.6
graphene-sqlalchemy==1.1.1
graphene==1.4.1
graphql-core==1.1
graphql-relay==0.4.5
hiredis==0.2.0
httplib2==0.11.1          # via google-api-python-client, oauth2client
humanfriendly==4.8        # via coloredlogs, executor, property-manager, rotate-backups
identify==1.0.8           # via pre-commit
idna==2.6                 # via cryptography
imagesize==1.0.0          # via sphinx
imaplib2==2.45.0
ipaddress==1.0.19         # via cryptography
ipython-genutils==0.2.0   # via traitlets
ipython==4.2.0
iso639==0.1.4
iso8601==0.1.12           # via colander, graphene-sqlalchemy
isodate==0.5.4
jdcal==1.3                # via openpyxl
jinja2==2.8
jsonpath-ng==1.4.3
kombu==4.2.1              # via celery
lingua==4.9
lxml==3.6.0
mako==1.0.7               # via alembic, pyramid-mako
markupsafe==1.0           # via jinja2, mako, pyramid-jinja2
mccabe==0.6.1             # via flake8
meld3==1.0.2              # via supervisor
mock==2.0.0
monotonic==1.4            # via fasteners, humanfriendly
montague==0.2.1           # via pyramid-autodoc
naturalsort==1.5.1        # via rotate-backups
ndg_httpsclient==0.4.1
nodeenv==1.3.0
oauth2client==4.1.2       # via google-api-python-client
oauthlib==2.0.7           # via requests-oauthlib, social-auth-core
openpyxl==2.4.1
paramiko==2.4.1           # via fabric
pastedeploy==1.5.2        # via pyramid
pathlib2==2.3.0           # via pickleshare
pbr==4.0.0                # via mock
pdbsublimetextsupport==0.2
pexpect==4.4.0            # via ipython
pickleshare==0.7.4        # via ipython
pip-tools==1.11.0
ply==3.11                 # via jsonpath-ng
polib==1.0.7
pre-commit==1.8.2
premailer==3.0.0
promise==2.1              # via graphene, graphql-core, graphql-relay
property-manager==2.2     # via executor, rotate-backups
psycopg2==2.7.3.1
ptyprocess==0.5.2         # via pexpect
py==1.5.3                 # via pytest
pyasn1-modules==0.2.1     # via oauth2client
pyasn1==0.4.2             # via oauth2client, paramiko, pyasn1-modules, rsa
pycodestyle==2.3.1        # via flake8
pycparser==2.18           # via cffi
pycrypto==2.6.1
pydot==1.2.4              # via sqlalchemy-schemadisplay
pyflakes==1.6.0           # via flake8
pygments==2.2.0           # via pyramid-debugtoolbar, sphinx
pyisemail==1.3.1
pyjwt==1.4.0
pylibmc==1.5.1
pynacl==1.2.1             # via paramiko
pyopenssl==17.5.0         # via ndg-httpsclient
pyparsing==2.2.0          # via pydot
pyramid-autodoc==1.0.0
pyramid-debugtoolbar-ajax==0.0.4
pyramid-debugtoolbar==3.0.1
pyramid-ipython==0.1
pyramid-jinja2==2.6.2
pyramid-mailer==0.14.1
pyramid-mako==1.0.2       # via pyramid-debugtoolbar
pyramid-tm==0.12.1
pyramid==1.7
pystemmer==1.3.0
pytest-cov==2.2.1
pytest-localserver==0.3.5
pytest-mock==1.1
pytest-splinter==1.7.3
pytest==2.9.2
python-dateutil==2.5.3
python-editor==1.0.3      # via alembic
python-json-logger==0.1.8
python-openid==2.2.5      # via social-auth-core
python-saml==2.2.0
pytz==2016.4
pyyaml==3.12              # via aspy.yaml, pre-commit
pyzmq==15.2.0
raven==6.9.0
redis==2.10.5
repoze.lru==0.7           # via pyramid, pyramid-debugtoolbar
repoze.sendmail==4.4.1    # via pyramid-mailer
requests-oauthlib==0.6.1
requests==2.10.0
rotate-backups==4.4
rsa==3.4.2                # via oauth2client
scandir==1.7              # via pathlib2
selenium==3.11.0          # via pytest-splinter, splinter
simpleeval==0.9.1         # via rotate-backups
simplegeneric==0.8.1      # via ipython
simplejson==3.8.2
singledispatch==3.4.0.3   # via graphene-sqlalchemy, tornado
six==1.10.0
snowballstemmer==1.2.1    # via sphinx
social-auth-app-pyramid==1.0.0
social-auth-core==1.7.0
social-auth-storage-sqlalchemy==1.1.0
sphinx==1.5.1
sphinxcontrib-httpdomain==1.5.0
sphinxcontrib-zopeext==0.2.1
splinter==0.7.6
sqlalchemy-schemadisplay==1.3
sqlalchemy==1.2.2
sqlparse==0.2.4
structlog==17.2.0
supervisor==3.3.0
termcolor==1.1.0
tornado==4.3
traitlets==4.3.2          # via ipython
transaction==1.6.1
translationstring==1.3
typing==3.6.4             # via promise
unidecode==0.4.19
uritemplate==3.0.0        # via google-api-python-client
urllib3==1.22             # via elasticsearch
uwsgitop==0.9
venusian==1.1.0           # via pyramid
verboselogs==1.7
vine==1.1.4               # via amqp
virtualenv==15.2.0        # via pre-commit
waitress==1.1.0           # via webtest
webob==1.6.1
webtest==2.0.21
werkzeug==0.11.10
zope.component==4.2.2
zope.deprecation==4.3.0   # via pyramid, pyramid-jinja2
zope.event==4.3.0         # via zope.component
zope.interface==4.2.0
zope.sqlalchemy==0.7.7
